remove the schemastore check and make it post publish recommendations
