+++ scripts/release/release.js
@@ -27,7 +27,7 @@ async function run() {
     require("./steps/push-to-git"),
     require("./steps/publish-to-npm"),
     require("./steps/bump-prettier"),
-    require("./steps/check-schema-store")
+    require("./steps/post-publish-steps")
   ];
 
   try {
